# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to the HavenStay Backend API documentation! 🚀\n\nHavenStay is a modern, high-performance Domain-Driven Design (DDD) backend built in Go. It empowers Hosts to manage luxury properties and Guests to seamlessly book remote getaways.\n\n### Key Features:\n- **🏘️ Property Management**: Full CRUD lifecycle for short-term rentals and luxury homes.\n- **👩‍💼 Host \u0026 Guest Profiles**: Independent onboarding and profile management domains.\n- **📅 Booking Engine**: Real-time reservation creation.\n- **❤️ Wishlists**: Guests can curate lists of favorite properties.\n\n**Built With**: Go 1.21+, Gin (REST), gRPC, Protobufs, MySQL 8, GORM.",
        "title": "HavenStay - Hotel \u0026 Room Booking API",
        "termsOfService": "https://havenstay.localhost/terms",
        "contact": {
            "name": "HavenStay Platform Support",
            "url": "https://github.com/AshutoshKY/havenstay-backend",
            "email": "api-support@havenstay.local"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/bookings": {
            "get": {
                "description": "Poll the history of all active, past, and cancelled reservations tied to a specific Guest profile.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a Guest's Booking History",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The ID of the Guest requesting their itinerary.",
                        "name": "guestId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of all historical and upcoming bookings for the Guest.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.ListBookingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing or invalid guest ID.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Securely initiates a booking transaction for a specific Property by a Guest.\nThe system will evaluate the requested dates, calculate total pricing, and attempt to reserve the asset.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a Booking Reservation",
                "parameters": [
                    {
                        "description": "Payload specifying the Guest ID, Property ID, Start Date, End Date, and calculated total price.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Reservation successfully created and locked.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.BookingResponse"
                        }
                    },
                    "400": {
                        "description": "Validation Bad Request - Overlapping dates or missing required fields.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/bookings/{id}": {
            "get": {
                "description": "Fetch an exact copy of a completed or pending Booking reservation by its unique confirmation ID.\nTypical use-case: Displaying a digital receipt or itinerary details to the Guest.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Retrieve Booking Invoice",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The numeric Reservation/Booking ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The retrieved itinerary and invoice details.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.BookingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid ID format.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - The reservation does not exist.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/guests": {
            "post": {
                "description": "Creates a new Guest profile. Guests are users who can book properties and create wishlists.\nEnsure the email address provided is unique across the platform.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Register a new Guest",
                "parameters": [
                    {
                        "description": "Payload containing the Guest's personal details (Name, Email, Phone).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.CreateGuestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Guest successfully registered and assigned a unique ID.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.GuestResponse"
                        }
                    },
                    "400": {
                        "description": "Validation Error - Invalid payload format or duplicate email.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/guests/{id}": {
            "get": {
                "description": "Fetches the profile and identity data for a specific Guest by their unique numeric identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guests"
                ],
                "summary": "Retrieve Guest details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The unique numeric ID of the Guest to retrieve.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested Guest details.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.GuestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The provided ID was invalid.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - No Guest exists with the provided ID.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_user_airbnb-test_api_proto_v1.Booking": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "end_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "guest_id": {
                    "type": "integer"
                },
                "payment_info": {
                    "type": "string"
                },
                "property_id": {
                    "type": "integer"
                },
                "start_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "status": {
                    "description": "e.g., PENDING, CONFIRMED, CANCELLED",
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.BookingResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Booking"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.CreateBookingRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "guest_id": {
                    "type": "integer"
                },
                "payment_info": {
                    "type": "string"
                },
                "property_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.CreateGuestRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.Guest": {
            "type": "object",
            "properties": {
                "booking_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "guest_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.GuestResponse": {
            "type": "object",
            "properties": {
                "guest": {
                    "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Guest"
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.ListBookingsResponse": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Booking"
                    }
                }
            }
        },
...
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. This field is\nthe nanosecond portion of the duration, not an alternative to seconds.\nNegative second values with fractions must still have non-negative nanos\nvalues that count forward in time. Must be between 0 and 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must\nbe between -315576000000 and 315576000000 inclusive (which corresponds to\n0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).",
                    "type": "integer"
                }
            }
        }
    }
}